Warehouse api and field backend validation

+++ src/main/java/com/whydigit/efit/repo/WarehouseRepository.java
@@ -7,7 +7,7 @@ import org.springframework.data.jpa.repository.Query;
 
 import com.whydigit.efit.entity.WarehouseVO;
 
-public interface WarehouseRepository extends JpaRepository<WarehouseVO, Integer>{
+public interface WarehouseRepository extends JpaRepository<WarehouseVO, Long>{
 @Query(value = "select w.warehouseLocation  from WarehouseVO w where w.orgId=?1")
 	Set<Object[]> getWarehouseLocationByOrgID(Long orgId);
